+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 OpenStudio Analysis Gem Change Log
 ==================================
 
+Version 1.0.0.rc10
+------------------
+* Fix boolean data type in datapoints translator
+* Allow __skip__ variable in datapoints translator
+
 Version 1.0.0.rc9
 -----------------
 * Fix bug in batch datapoints to look for outputs_json, not outputs when importing the definition of the outputs JSON file.